+++ DecaTec.WebDav/WebDavSessionItem.cs
@@ -615,6 +615,7 @@ namespace DecaTec.WebDav
         /// <summary>
         /// Gets the <see cref="AdditionalWebDavProperties"/> representing the additional WebDAV properties not defined in <see href="https://www.ietf.org/rfc/rfc4918.txt">RFC 4918</see>, <see href="https://tools.ietf.org/html/rfc4331">RFC 4331</see>, <see href="https://tools.ietf.org/html/draft-hopmann-collection-props-00">Additional WebDAV Collection Properties</see> or the IIS WebDAV specification.
         /// </summary>
+        /// <remarks>Important: Additional WebDAV properties are currently not supported on Xamarin. See <see href="https://github.com/DecaTec/Portable-WebDAV-Library/wiki/Xamarin">the project's wiki</see> for more information about the Portable WebDAV Library used on Xamarin.</remarks>
         public AdditionalWebDavProperties AdditionalProperties
         {
             get